fs: build failure when compiling rocksdb in debug mode
When building with DEBUG_LEVEL=1 with below command,
"make clean;DEBUG_LEVEL=1 ROCKSDB_PLUGINS=zenfs make -j48 db_bench db_stress ldb"
compilation fails with error:
"error: ‘std::string rocksdb::GetLogFilename(std::string)’ defined but not used"

Current logic defines the function GetLogFileName() for all debug levels(0,1,2)
but uses it only when debug-level=0. So for debug builds it gets defined but
is not used, resulting in the error. Also logs are enabled in production
builds now. Related to 7f8392e.

Define it such that is enabled for debug-level > 0 OR when TERARKDB is defined.

Signed-off-by: Aravind Ramesh <Aravind.Ramesh@wdc.com>